Add test cases for empty children, also collaps tests
